requirements.txt
pandas==1.4.4
plotly==5.11.0
python_dateutil==2.8.2
streamlit==1.14.0
yfinance==0.1.85
